requirements.txt
streamlit>=1.30.0
groq>=0.4.0
streamlit>=1.29.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
torch>=2.0.0
